+++ 3.PresentationLayers/Views/Form_TheLoai.cs
@@ -15,12 +15,12 @@ using static System.ComponentModel.Design.ObjectSelectorEditor;
 
 namespace _3.PresentationLayers.Views
 {
-    public partial class FrmTheLoai : Form
+    public partial class Form_TheLoai : Form
     {
         ITheLoaiService _iTLService;
         TheLoai _tl;
         public Guid SelectID { get; set; }
-        public FrmTheLoai()
+        public Form_TheLoai()
         {
             InitializeComponent();
             _iTLService = new TheLoaiService();
@@ -42,6 +42,7 @@ namespace _3.PresentationLayers.Views
             {
                 dtg_Show.Rows.Add(stt++, x.Id, x.Ma, x.Ten, x.TrangThai == 0 ? "Không hoạt động" : "Hoạt động");
             }
+            dtg_Show.AllowUserToAddRows = false;
         }
         private void Load()
         {
@@ -64,10 +65,10 @@ namespace _3.PresentationLayers.Views
         {
             LoadDataToDtg(_iTLService.GetAll());
             _tl = null;
-            tb_ma.Text = "";
-            tb_ten.Text = "";
+            tb_ma.Text = string.Empty;
+            tb_ten.Text = string.Empty;
             cbb_trangthai.Text = "--Chọn--";
-            tb_Timkiem.Text = "";
+            tb_Timkiem.Text = string.Empty;
         }
         private TheLoai GetDataFromGui_Them()
         {
@@ -93,6 +94,7 @@ namespace _3.PresentationLayers.Views
         /*=============================================================================*/
         private void dtg_Show_CellClick(object sender, DataGridViewCellEventArgs e)
         {
+            if (e.RowIndex == -1) return;
             SelectID = Guid.Parse(dtg_Show.CurrentRow.Cells[1].Value.ToString());
             tb_ma.Text = dtg_Show.CurrentRow.Cells[2].Value.ToString();
             tb_ten.Text = dtg_Show.CurrentRow.Cells[3].Value.ToString();
@@ -100,20 +102,46 @@ namespace _3.PresentationLayers.Views
         }
         private void btn_Them_Click(object sender, EventArgs e)
         {
-            MessageBox.Show(_iTLService.Add(GetDataFromGui_Them()));
-            ResetForm();
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn thêm đối tượng không?", "Thông báo", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                if (tb_ma.Text == string.Empty || tb_ten.Text == string.Empty ||cbb_trangthai.Text=="--Chọn--")
+                {
+                    MessageBox.Show("Bạn chưa nhập đầy đủ thông tin", "Thông báo", MessageBoxButtons.OK);
+                    return;
+                }
+                MessageBox.Show(_iTLService.Add(GetDataFromGui_Them()));
+                ResetForm();
+            }
+            else return;
+
         }
 
         private void btn_Sua_Click(object sender, EventArgs e)
         {
-            MessageBox.Show(_iTLService.Update(GetDataFromGui_Sua_Xoa()));
-            ResetForm();
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn cập nhật đối tượng không?", "Thông báo", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                if (tb_ma.Text == string.Empty || tb_ten.Text == string.Empty || cbb_trangthai.Text == "--Chọn--")
+                {
+                    MessageBox.Show("Bạn chưa nhập đầy đủ thông tin", "Thông báo", MessageBoxButtons.OK);
+                    return;
+                }
+                MessageBox.Show(_iTLService.Update(GetDataFromGui_Sua_Xoa()));
+                ResetForm();
+            }
+            else return;
         }
 
         private void btn_Xoa_Click(object sender, EventArgs e)
         {
-            MessageBox.Show(_iTLService.Delete(GetDataFromGui_Sua_Xoa()));
-            ResetForm();
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn xóa đối tượng không?", "Thông báo", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                MessageBox.Show(_iTLService.Delete(GetDataFromGui_Sua_Xoa()));
+                ResetForm();
+            }
+            else return;
         }
 
         private void btn_Reset_Click(object sender, EventArgs e)
